# docs/api/blueiq-api-gateway-openapi.yaml
openapi: 3.0.3
info:
  title: BlueIQ API Gateway
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.example.com/support
    email: support@example.com
  description: |
    This document describes the API for the BlueIQ API Gateway.
    It serves as the single entry point for all client interactions with the BlueIQ microservices.
    This specification is designed to be a reference for consumers and a template for similar gateway patterns.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://dev.blueiq.example.com/api/v1
    description: Development environment
  - url: https://staging.blueiq.example.com/api/v1
    description: Staging environment
  - url: https://blueiq.example.com/api/v1
    description: Production environment
security:
  - BearerAuth: []
...
paths:
...
  /roles:
    get:
      tags:
        - Role Management
      summary: List all roles
      operationId: listRoles
      description: Retrieves a paginated list of all user roles. Supports query parameters for pagination.
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of roles per page
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: A list of roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Role Management
      summary: Create a new role
      operationId: createRole
      description: Creates a new user role. Requires a role name. Optional fields include description and a list of permissions.
      security:
        - BearerAuth: []
      requestBody:
        description: Role object to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          description: Role already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /roles/{roleId}:
    get:
      tags:
        - Role Management
      summary: Get a role by ID
      operationId: getRoleById
      description: Retrieves the details of a specific user role by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          description: ID of the role to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Role Management
      summary: Update an existing role
      operationId: updateRole
      description: Updates an existing user role. Allows modification of name, description, and permissions.
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          description: ID of the role to update
          schema:
            type: string
            format: uuid
      requestBody:
        description: Role object to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Role Management
      summary: Delete a role by ID
      operationId: deleteRole
      description: Deletes a specific user role by its unique ID. This operation may affect users assigned to this role.
      security:
        - BearerAuth: []
      parameters:
        - name: roleId
          in: path
          required: true
          description: ID of the role to delete
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Invalid input provided.
        details:
          type: string
          nullable: true
          example: The 'username' field is required.
...
    Pagination:
      type: object
      properties:
        totalItems:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        currentPage:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
...
    Role:
      type: object
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          example: f47ac10b-58cc-4372-a567-0e02b2c3d479
          readOnly: true
        name:
          type: string
          example: Administrator
          description: The name of the role.
        description:
          type: string
          example: Has full system access.
          nullable: true
          description: A description of the role.
        permissions:
          type: array
          items:
            type: string
          example: ["users:create", "users:read", "users:update", "users:delete"]
          description: List of permissions associated with the role.
        createdAt:
          type: string
          format: date-time
          example: '2023-01-10T09:00:00Z'
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          example: '2023-01-10T09:00:00Z'
          readOnly: true
    RoleListResponse:
      type: object
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        pagination:
          $ref: '#/components/schemas/Pagination'
    CreateRoleRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Editor
        description:
          type: string
          example: Can edit content.
          nullable: true
        permissions:
          type: array
          items:
            type: string
          example: ["content:edit", "content:publish"]
          description: List of permissions for the new role.
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          example: Senior Editor
        description:
          type: string
          example: Can edit and approve content.
          nullable: true
        permissions:
          type: array
          items:
            type: string
          example: ["content:edit", "content:publish", "content:approve"]
          description: List of permissions for the role. This will replace existing permissions.
...
  responses:
    UnauthorizedError:
      description: Unauthorized - Authentication token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing token
              value:
                statusCode: 401
                message: "Authentication token is missing."
            invalidToken:
              summary: Invalid token
              value:
                statusCode: 401
                message: "Authentication token is invalid or expired."
    ForbiddenError:
      description: Forbidden - User does not have permission to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Insufficient permissions
              value:
                statusCode: 403
                message: "You do not have the necessary permissions to perform this action."
    NotFoundError:
      description: Not Found - The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            resourceNotFound:
              summary: Resource not found
              value:
                statusCode: 404
                message: "The requested resource was not found."
                details: "User with ID '12345' does not exist."
...
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            genericError:
              summary: Generic server error
              value:
                statusCode: 500
                message: "An unexpected error occurred. Please try again later or contact support."
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token for authentication. The token should be included in the
        Authorization header as 'Bearer {token}'.
        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`