# generated/aspnetcore-server/src/BlueIQ.Api.Server/Controllers/RoleManagementApi.cs
/*
 * BlueIQ API Gateway
 *
 * This document describes the API for the BlueIQ API Gateway. It serves as the single entry point for all client interactions with the BlueIQ microservices. This specification is designed to be a reference for consumers and a template for similar gateway patterns. 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@example.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using BlueIQ.Api.Server.Attributes;
using BlueIQ.Api.Server.Models;

namespace BlueIQ.Api.Server.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class RoleManagementApiController : ControllerBase
    { 
        /// <summary>
        /// Create a new role
        /// </summary>
        /// <remarks>Creates a new user role. Requires a role name. Optional fields include description and a list of permissions.</remarks>
        /// <param name="createRoleRequest">Role object to be created</param>
        /// <response code="201">Role created successfully</response>
        /// <response code="400">Invalid input</response>
        /// <response code="401">Unauthorized - Authentication token is missing or invalid.</response>
        /// <response code="403">Forbidden - User does not have permission to access this resource.</response>
        /// <response code="409">Role already exists</response>
        /// <response code="500">Internal Server Error - An unexpected error occurred on the server.</response>
        [HttpPost]
        [Route("/api/v1/roles")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreateRole")]
        [SwaggerResponse(statusCode: 201, type: typeof(Role), description: "Role created successfully")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "Invalid input")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Unauthorized - Authentication token is missing or invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Forbidden - User does not have permission to access this resource.")]
        [SwaggerResponse(statusCode: 409, type: typeof(ErrorResponse), description: "Role already exists")]
        [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponse), description: "Internal Server Error - An unexpected error occurred on the server.")]
        public virtual IActionResult CreateRole([FromBody]CreateRoleRequest createRoleRequest)
        {

            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default);
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default);
            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409, default);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"createdAt\" : \"2023-01-10T09:00:00Z\",\r\n  \"permissions\" : [ \"users:create\", \"users:read\", \"users:update\", \"users:delete\" ],\r\n  \"name\" : \"Administrator\",\r\n  \"description\" : \"Has full system access.\",\r\n  \"id\" : \"f47ac10b-58cc-4372-a567-0e02b2c3d479\",\r\n  \"updatedAt\" : \"2023-01-10T09:00:00Z\"\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete a role by ID
        /// </summary>
        /// <remarks>Deletes a specific user role by its unique ID. This operation may affect users assigned to this role.</remarks>
        /// <param name="roleId">ID of the role to delete</param>
        /// <response code="204">Role deleted successfully</response>
        /// <response code="401">Unauthorized - Authentication token is missing or invalid.</response>
        /// <response code="403">Forbidden - User does not have permission to access this resource.</response>
        /// <response code="404">Not Found - The requested resource could not be found.</response>
        /// <response code="500">Internal Server Error - An unexpected error occurred on the server.</response>
        [HttpDelete]
        [Route("/api/v1/roles/{roleId}")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("DeleteRole")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Unauthorized - Authentication token is missing or invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Forbidden - User does not have permission to access this resource.")]
        [SwaggerResponse(statusCode: 404, type: typeof(ErrorResponse), description: "Not Found - The requested resource could not be found.")]
        [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponse), description: "Internal Server Error - An unexpected error occurred on the server.")]
        public virtual IActionResult DeleteRole([FromRoute (Name = "roleId")][Required]Guid roleId)
        {

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get a role by ID
        /// </summary>
        /// <remarks>Retrieves the details of a specific user role by its unique ID.</remarks>
        /// <param name="roleId">ID of the role to retrieve</param>
        /// <response code="200">Successful operation</response>
        /// <response code="401">Unauthorized - Authentication token is missing or invalid.</response>
        /// <response code="403">Forbidden - User does not have permission to access this resource.</response>
        /// <response code="404">Not Found - The requested resource could not be found.</response>
        /// <response code="500">Internal Server Error - An unexpected error occurred on the server.</response>
        [HttpGet]
        [Route("/api/v1/roles/{roleId}")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("GetRoleById")]
        [SwaggerResponse(statusCode: 200, type: typeof(Role), description: "Successful operation")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Unauthorized - Authentication token is missing or invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Forbidden - User does not have permission to access this resource.")]
        [SwaggerResponse(statusCode: 404, type: typeof(ErrorResponse), description: "Not Found - The requested resource could not be found.")]
        [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponse), description: "Internal Server Error - An unexpected error occurred on the server.")]
        public virtual IActionResult GetRoleById([FromRoute (Name = "roleId")][Required]Guid roleId)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"createdAt\" : \"2023-01-10T09:00:00Z\",\r\n  \"permissions\" : [ \"users:create\", \"users:read\", \"users:update\", \"users:delete\" ],\r\n  \"name\" : \"Administrator\",\r\n  \"description\" : \"Has full system access.\",\r\n  \"id\" : \"f47ac10b-58cc-4372-a567-0e02b2c3d479\",\r\n  \"updatedAt\" : \"2023-01-10T09:00:00Z\"\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// List all roles
        /// </summary>
        /// <remarks>Retrieves a paginated list of all user roles. Supports query parameters for pagination.</remarks>
        /// <param name="page">Page number for pagination</param>
        /// <param name="pageSize">Number of roles per page</param>
        /// <response code="200">A list of roles</response>
        /// <response code="401">Unauthorized - Authentication token is missing or invalid.</response>
        /// <response code="403">Forbidden - User does not have permission to access this resource.</response>
        /// <response code="500">Internal Server Error - An unexpected error occurred on the server.</response>
        [HttpGet]
        [Route("/api/v1/roles")]
        [Authorize]
        [ValidateModelState]
        [SwaggerOperation("ListRoles")]
        [SwaggerResponse(statusCode: 200, type: typeof(RoleListResponse), description: "A list of roles")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Unauthorized - Authentication token is missing or invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Forbidden - User does not have permission to access this resource.")]
        [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponse), description: "Internal Server Error - An unexpected error occurred on the server.")]
        public virtual IActionResult ListRoles([FromQuery (Name = "page")]int? page, [FromQuery (Name = "pageSize")]int? pageSize)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"pagination\" : {\r\n    \"totalItems\" : 100,\r\n    \"totalPages\" : 5,\r\n    \"pageSize\" : 20,\r\n    \"currentPage\" : 1\r\n  },\r\n  \"roles\" : [ {\r\n    \"createdAt\" : \"2023-01-10T09:00:00Z\",\r\n    \"permissions\" : [ \"users:create\", \"users:read\", \"users:update\", \"users:delete\" ],\r\n    \"name\" : \"Administrator\",\r\n    \"description\" : \"Has full system access.\",\r\n    \"id\" : \"f47ac10b-58cc-4372-a567-0e02b2c3d479\",\r\n    \"updatedAt\" : \"2023-01-10T09:00:00Z\"\r\n  }, {\r\n    \"createdAt\" : \"2023-01-10T09:00:00Z\",\r\n    \"permissions\" : [ \"users:create\", \"users:read\", \"users:update\", \"users:delete\" ],\r\n    \"name\" : \"Administrator\",\r\n    \"description\" : \"Has full system access.\",\r\n    \"id\" : \"f47ac10b-58cc-4372-a567-0e02b2c3d479\",\r\n    \"updatedAt\" : \"2023-01-10T09:00:00Z\"\r\n  } ]\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<RoleListResponse>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Update an existing role
        /// </summary>
        /// <remarks>Updates an existing user role. Allows modification of name, description, and permissions.</remarks>
        /// <param name="roleId">ID of the role to update</param>
        /// <param name="updateRoleRequest">Role object to be updated</param>
        /// <response code="200">Role updated successfully</response>
        /// <response code="400">Invalid input</response>
        /// <response code="401">Unauthorized - Authentication token is missing or invalid.</response>
        /// <response code="403">Forbidden - User does not have permission to access this resource.</response>
        /// <response code="404">Not Found - The requested resource could not be found.</response>
        /// <response code="500">Internal Server Error - An unexpected error occurred on the server.</response>
        [HttpPut]
        [Route("/api/v1/roles/{roleId}")]
        [Authorize]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("UpdateRole")]
        [SwaggerResponse(statusCode: 200, type: typeof(Role), description: "Role updated successfully")]
        [SwaggerResponse(statusCode: 400, type: typeof(ErrorResponse), description: "Invalid input")]
        [SwaggerResponse(statusCode: 401, type: typeof(ErrorResponse), description: "Unauthorized - Authentication token is missing or invalid.")]
        [SwaggerResponse(statusCode: 403, type: typeof(ErrorResponse), description: "Forbidden - User does not have permission to access this resource.")]
        [SwaggerResponse(statusCode: 404, type: typeof(ErrorResponse), description: "Not Found - The requested resource could not be found.")]
        [SwaggerResponse(statusCode: 500, type: typeof(ErrorResponse), description: "Internal Server Error - An unexpected error occurred on the server.")]
        public virtual IActionResult UpdateRole([FromRoute (Name = "roleId")][Required]Guid roleId, [FromBody]UpdateRoleRequest updateRoleRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default);
            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default);
            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default);
            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default);
            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default);
            string exampleJson = null;
            exampleJson = "{\r\n  \"createdAt\" : \"2023-01-10T09:00:00Z\",\r\n  \"permissions\" : [ \"users:create\", \"users:read\", \"users:update\", \"users:delete\" ],\r\n  \"name\" : \"Administrator\",\r\n  \"description\" : \"Has full system access.\",\r\n  \"id\" : \"f47ac10b-58cc-4372-a567-0e02b2c3d479\",\r\n  \"updatedAt\" : \"2023-01-10T09:00:00Z\"\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            exampleJson = "{\r\n  \"details\" : \"The 'username' field is required.\",\r\n  \"message\" : \"Invalid input provided.\",\r\n  \"statusCode\" : 400\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<Role>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}